latihan-fastapi/requirements.txt
alembic
bcrypt
couchbase
databases
datetime
fastapi[all]
gunicorn
kafka-python
pydantic
passlib
psycopg2-binary
python-jose[pycryptodome]
python-multipart
PyJWT
sqlalchemy
typing
tk
pytho
uuid
uvicorn